Updated program to help with an entire Wordle game
The program now assists with an entire Wordle game involving several guesses, and not just a single guess.

KNOWN ISSUE: The program still suggests some words that contain letters known to not be in the word. This doesn't impede functionality too much, but I'll still be looking into how to correct it.